dnstogrpc: Implement systemd socket activation
This patch makes dnstogrpc able to get sockets from systemd.

It's very useful for listening on port 53 without having to use problematic
tricks like capabilities or iptables redirects.

For grpctodns this is less useful as it doesn't need privileged ports, but it
may be implemented in the future anyway for completion.