dnstox: Implement fallback lookups
If dnss is the resolver for the machine it is running on, we can run into
problems resolving the servers we proxy to (like "dns.google.com." in
dns-over-https mode), due to a circular dependency.

To solve that problem, this patch introduces fallback lookups: there is a
command-line controlled list of domains for which we bypass the proxying, and
resolve directly using DNS to the given upstream server.

This is enabled by default for "dns.google.com.", to simplify the
configuration of the dns-over-https mode.