Use glog.Fatal when exiting due to fatal errors
On severe errors, specially when initializing, we want to make sure we exit
with a failure exit code.

Currently, the code is not consistent with this, we use glog.Error and then
sometimes we panic, sometimes we don't.

This patch makes the fatal error handling more consistent by using glog.Fatal
in all cases.