dnstox: Return more accurate TTLs
This patch makes the cache resolver keep and return more accurate TTLs.

Before, we used to always return the minimum TTL to clients, which was not too
bad but caused a bit of waste, and introduced noise if dnss was not the last
server in the chain.

With this change, dnss will return the real TTL (accurate to within our
maintenance period = 30s), capped to prevent really long entries for
monopolizing the cache.