dnstox: Set the question explicitly on the httpsResolver's reply
The current httpsResolver uses dns.Msg.SetQuestion to set the question on the
reply, but it's wrong as the code also sets the id, and sets the recursion
desired bit (it's intended to construct requests, not replies).

This patch fixes that by setting the question directly.

Thanks to Rodrigo Campos for the bug report!