dnstox: Remove answer race in the caching resolver
The cache maintenance goroutine regularly modifies the cached answers,
updating their TTLs.

When responding to queries, we use those cached answers, but currently there
is a race because the answers we return could be modified by the maintenance
goroutine concurrently.

This patch fixes that race by never modifying the cached answers, but updating
copies while doing maintenance. This is much faster than copying on each cache
hit, but may extend the maintenance period.