dnstox: Remove race in tests
One of the tests has a data race, as it changes the maintenance period
"constant". This patch fixes the race by starting maintenance only once, after
the period has been updated.

It's not a pretty approach in general, but in this case, as we only test the
function once, it works fine.