Track dependencies with glide (experiment)
glide is a Go tool for vendoring, which we will use here to make it easier to
track dependencies and versions that we have tested against.

Users are not expected to use it, dnss should just build and work.

But using glide for development means it's easier to track the versions we've
tested against, and identify when backwards-incompatible dependency changes
break dnss. This is an unfortunate consequence of how golang does
dependencies.

Note we intentionally do not include all the dependencies in the repository.

This is an experiment, and may be reverted or changed if it doesn't work, so
don't rely on it for packaging or automation.