Use a more standard testing layout
The "testing" directory is awkward, not very common, and it makes some
things more difficult.

This patch moves the tests next to the main package they are testing and
the utilities to the internal directory, leaving the "testing" directory
empty.

The end result is that the tests are now organized more like any other
Go program.