Replace github.com/golang/glog with blitiri.com.ar/go/log
For the most common uses of dnss, using github.com/golang/glog
introduces a bit of complexity as it needs to be flushed regularly and
before exit, and also requires some parameters to be nice to systemd and
log rotators.

This patch replaces it with blitiri.com.ar/go/log, which has a very
similar API and functionality, but it does not require flushing and has
more practical defaults.