dnsserver: Propagate errors from the resolver
When dnsserver gets an error from the resolver, it does not reply to the
user back, resulting in client-side timeouts.

This patch fixes this problem by making the server reply with SERVFAIL
when the resolver returned errors.

This matches the behaviour we already use for unqualified and fallback
queries.

A series of tests for these behaviours are added to prevent future
regressions.