tests: Add coverage and integration test scripts
This patch adds scripts for integration tests against external services,
and to generate code coverage reports.

Note the script for testing against external services is brittle and is
not hermetic, so it is prone to false negatives. However, it is a
valuable tool to check that dnss can inter-operate with common public
providers.