httpresolver: Limit message sizes to 64k
In order to reduce the risk of denial of service and similar issues, we
want to limit the answer message sizes. Today, the limit for DoH is 4k,
and there is no limit for the JSON protocol.

There's been discussions in the DoH mailing list about increasing the
maximum message size to 64k, so this patch applies that limit to both
protocols, to err on the side of caution.