dnss: Use httpproxy to find the proxy server domain
To extract the proxy server domain, this patch switches from using
the built-in http to the external httpproxy package.

The former's (new) built-in caching makes testing more difficult, and
the httpproxy package is part of golang.org/x/net so it is not a
new dependency.

As a part of this change we get rid of extractHostname, which was
a temporary hack and is no longer needed (as we don't support Go < 1.9
anymore).