tests: Speed up the tests by waiting a little before querying over UDP
When we query over UDP, we don't wait for the server to come up, which
means the first request is usually lost, and we need to wait for it to
time out and then retry. This is not a big issue but it slows down the
tests a few seconds.

This patch makes the tests wait 200ms before issuing the query, which is
often long enough for the server to come up.

There is no difference in the result, this only causes the tests to be
faster in the common case.