travis: Work around network reachability issues
Travis has a networking issue where it can't reach blitiri.com.ar, which
causes the build to fail on Go 1.11.

Go >= 1.12 work because it uses the Go Proxy instead of contacting the
host directly, which is a horrible workaround for the issue, but it's
effective.

Until the problem is solved, mark the Go 1.11 as optional.