Rename -testing__insecure_http to -insecure_http_server
The -testing__insecure_http flag makes the HTTP server listen direcly on
HTTP instead of HTTPS. Originally, the flag was named to prevent
accidental misuse.

But there is a very reasonable use case for plain HTTP, where dnss is
deployed behind a reverse proxy. In that case, the flag name can be
confusing and misleading.

This was reported in https://github.com/albertito/dnss/issues/8.

This patch renames the testing flag to -insecure_http_server, which
should make its purpose more clear.