Remove DNS over JSON support
Now that DoH has gained significant adoption, there is not much reason
to keep support for DNS over JSON, which was never standarized.

This patch removes the code necessary for supporting DNS over JSON in
both client and server.

This is unfortunately a backwards-incompatible change for those that
were forcing the use of this protocol to contact a JSON-only server.
Those users will need to adjust how they run dnss, and use DoH.
I don't expect there are many people in that situation, though.