resolver: Fix custom CA pool file
The httpresolver supports receiving a custom file to use as root CA set,
via the --https_client_cafile option.

Unfortunately, the code handling this was configuring the TLS transport
incorrectly, and this file wasn't used as root CA.

So in practice the system CA set was used even if this option was passed.

This patch fixes this issue by setting the right internal option, and
also adds tests for this case.