httpresolver: React faster to network errors
When the TCP connection to the upstream server dies unexpectedly, dnss
can return errors for a very long time (hours) instead of reconnecting.
This is because the HTTP/2 transport will insist on using the existing
(dead) connection, as it doesn't give up on it.

This can easily happen when the computer dnss changes wifi networks, for
example.

Unfortunately, there is no practical way to tell the transport to drop
the connections.

So this patch introduces a workaround, which is to switch to a new
client when we've seen errors consistently for 10s. That's enough time
for 2 queries back to back to timeout (each query has a 4s timeout).