httpresolver: Always use the fallback resolver
When the httpresolver needs to resolve the name of our upstream server
(e.g. `dns.google.`), it uses the system resolver.

Normally the system resover is either configured to work fine
independently, or will point back to dnss, in which case dnss will use
`-fallback_domains` to help identify these recursive requests and use
`-fallback_upstream` to get an answer.

However, in some cases, the system resolver won't work properly or send
some requests to dnss that don't fall within `-fallback_domains`. An
example of this was reported in
https://github.com/albertito/dnss/issues/9, where `systemd-resolved`
does some DNSSEC resolutions to `.` and `google.` causing it to
mis-behave.

This patch makes dnss always use the fallback resolver, without falling
back to the system resolver at all. This should result in more
predictible behaviour and simpler setups, as now dnss should be fully
independent from the system resolver.

Thanks to David Mandelberg (david@mandelberg.org) for finding and
helping debug this issue.