tests: Create the directory for benchmark results
If the directory we use to store benchmark results for later comparison
doesn't exist, we need to create it. Otherwise, the run will fail.

A specific behaviour in git status caused this not to be caught before
by the automated tests (where the directory doesn't exist), because the
tree was always accidentally detected as dirty.