dnsserver: Truncate UDP responses that are too large
When an UDP response is larger than accepted by the client (defaults to
512, but can be extended using EDNS0), the standard says we should
truncate it.

Currently, we don't. In practice, this often works anyways because most
client libraries don't mind, but it can cause problems if the reply
exceeds the UDP packet size, and/or if the client is strict in checking
(like some Go libraries are).

This patch makes dnss conform to the standard by truncating DNS
responses over UDP as needed.

In those cases, the client is expected to retry over TCP (which doesn't
have this issue).

Note that unfortunately cleanbrowsing.org DoH resolver has to be removed
from the tests, because they don't handle this case correctly, and fail
all queries with large responses.