httpresolver: Make invalid server test even more hermetic
Currently, the invalid server test contacts 0.0.0.0, which unfortunately
is taken as localhost:80, and if there is a valid web server running
there, the test fails.

This patch fixes the problem by creating a test server, closing it, and
then attempting to contact it. That should be much more hermetic and
independent from the testing environment.