modules: Update dependencies' versions
This patch updates the versions of our dependencies, just to keep up
with general bug fixes. No API changes or compatibility issues are
expected.

The minimum supported Go version is raised to 1.18, which is (at the
point of this commit), the previous release, and already included in all
major Linux distributions (including Debian stable backports).