tests: Increase retries on the external tests
The external tests are heavily dependent on the testing environment,
because they reach out to live external resolvers.

Because of it, they are quite sensitive to transient failures, and this
is often experienced as flakiness in the Github CI.

To make these less flaky, this patch increases the retries, with
incremental delay between each attempt.